Extract AgentStateChart into reusable component
Move the Agent State Distribution pie chart from inline
Dashboard code into a shared component and reuse it on
the Agents page, keeping a single source of truth.
